chore: rename sink to store
